More tests passing
