+++ hapi-fhir-storage-batch2/src/main/java/ca/uhn/fhir/batch2/model/BaseWorkChunkEvent.java
@@ -1,7 +1,10 @@
 package ca.uhn.fhir.batch2.model;
 
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+
 public abstract class BaseWorkChunkEvent {
-	private String myChunkId;
+	protected final String myChunkId;
 
 	protected BaseWorkChunkEvent(String theChunkId) {
 		myChunkId = theChunkId;
@@ -11,7 +14,19 @@ public abstract class BaseWorkChunkEvent {
 		return myChunkId;
 	}
 
-	public void setChunkId(String theChunkId) {
-		myChunkId = theChunkId;
+	@Override
+	public boolean equals(Object theO) {
+		if (this == theO) return true;
+
+		if (theO == null || getClass() != theO.getClass()) return false;
+
+		BaseWorkChunkEvent that = (BaseWorkChunkEvent) theO;
+
+		return new EqualsBuilder().append(myChunkId, that.myChunkId).isEquals();
+	}
+
+	@Override
+	public int hashCode() {
+		return new HashCodeBuilder(17, 37).append(myChunkId).toHashCode();
 	}
 }
